Create README.md
Leveraging the power of Tableau, this dashboard provides a detailed view of critical metrics that influence our talent management and organizational strategies.

The line chart captures the trend of recruitment over time, allowing stakeholders to identify patterns and fluctuations. Staff levels in the Learning and Development (L&D) data are color-coded for a focused analysis of recruitment dynamics across different levels.

The gender count chart visualizes the distribution of employees by gender, offering a clear picture of gender diversity within the workforce.

The treemap offers an engaging visual representation of the distribution of employees based on their highest qualifications. This insight informs talent development and recruitment strategies.

Using bar charts break down the sources of employee recruitment and their last companies worked, providing valuable insights into the effectiveness of different recruitment channels and the professional backgrounds of the workforce.

Scatter plot investigates the potential influence of the number of companies worked prior to joining on an employee's median salary. The visual representation aids in identifying correlations.

This HR metrics dashboard empowers stakeholders with actionable insights, enabling data-driven decisions that contribute to our organization's success and sustainability.